requirements.txt
alembic==1.4.2
asgiref==3.2.10
beautifulsoup4==4.9.1
blinker==1.4
bootstrap4==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap4==1.1.1
django-crispy-forms==1.9.1
django-decouple==2.1
django-heroku==0.3.1
django-registration==2.4.1
django-rest-framework==0.1.0
django-tinymce==3.0.2
djangorestframework==3.11.0
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.1
pyuploadcare==2.7.0
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
whitenoise==5.1.0
WTForms==2.3.1
